display oauth2 error message from the server
Reviewed-on: https://gerrit.instructure.com/74219
Tested-by: Jenkins
Product-Review: Cody Cutrer <cody@instructure.com>
QA-Review: Cody Cutrer <cody@instructure.com>

+++ app/controllers/login/oauth2_controller.rb
@@ -27,6 +27,11 @@ class Login::Oauth2Controller < Login::OauthBaseController
   def create
     reset_session_for_login
 
+    if params[:error_description]
+      flash[:delegated_message] = Sanitize.clean(params[:error_description])
+      return redirect_to login_url
+    end
+
     if jwt['nonce'] != session.delete(:oauth2_nonce)
       raise ActionController::InvalidAuthenticityToken
     end
